Move gtxkey command to the key folder
